ninjabackend: include module requested symbols from static libraries
Even though shared modules are more or less incompatible with statically
linked binaries, it is of course possible to link static libraries into
a binary that uses shared modules.

Whenever an executable is statically linked with a library, the linker
omits library functions that are not used by the executable.  However,
this could include functions that are not used by the executable, but
are needed by shared modules.  When this happens, the shared module will
fail to load.  One common workaround in that case is to use link_whole,
but that is not always possible and might seriously bloat the executable.

This commit implements an alternative solution, inspired by the
operation of QEMU's Makefiles.  Module object files are processed
(for example with "nm") to find undefined symbols, and the undefined
symbols are included in the executable through the linker's "-u" option.
(There exists equivalents in the MSVC world for both steps, even though
they are not implemented yet).

This commit adds the logic to the generic code and to the ninja backend.
The next commit is where we actually start to use "nm".

Note that, unlike symbolextractor, this is not just an optimization.
However, if for example nm is not found in the toolchain file for
cross-compilation, behavior is unchanged compared to before this series:
a module that has dependencies on a statically-linked library will
fail to link, but everything else will work normally.