build: allow files and file-like objects as arguments to generators
The same file can sometimes be used many times in a generator's command
line, and it is unwieldy to always add it to the process() command line.
If the file is in the source directory, this is merely an annoyance: it
can be placed it in the arguments with current_source_dir(), even though
that is a hack and the generator is not rerun when the file changes.
Likewise if the file comes from configure_file.

If it is produced by a custom_target, however, placing it in the
arguments does not work because you do not get the right dependencies
and the generator may run too soon.

Fix all cases by allowing files and file-like objects as arguments to
a generator.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>