mtest: make _run_subprocess a context manager
Through an "async with" statement, SingleTestRunner will be able to run
code before _run_subprocess waits on the process.  For now all it does
is stashing away a tuple-like object that replaces the former return
value of _run_subprocess.  In the next patch, SingleTestRunner will
also get hold of the stdout and stderr StreamReaders.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>