ninjabackend: avoid useless isinstance calls
For QEMU, all the isinstance call sum up to about half a second
of execution time.  We know if something is a list or a single
dependency, so split add_dep and add_orderdep into two methods
depending on what they're called with---just like Python does
not do any magic and separates append from extend.