coredata: only warn about built-in option deprecation if meson_version is new enough
When an old value of a built-in option is deprecated, we would like to warn about it,
but at the same time not warn for all projects that possibly could not know about
the deprecation.  So pass the meson_version from project() down to coredata,
and use it to make the deprecation message conditional.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>