ninjabackend: unify building rpath args
Implement RustCompiler.build_rpath_args, so that more code can
be shared between non-Rust and Rust targets.  The new function
takes care of converting the arguments to "-C link-arg=" and
adding the rustup sysroot.  However, I'm not sure about the
latter code and I am just copying it as is, because it seems to add
the sysroot multiple times rather than just once.