vcs_tag: try using "git rev-parse --git-dir" to check if in a repo
If running from a git worktree, .git is not a directory but a
file.  In that case, "git rev-parse --git-dir" can be used to
check whether the directory is a worktree or just a place with
a ".git" file in it.

While at it, improve slightly the detection for non-worktree
git repos by checking also for a HEAD file.  This is a necessary
condition for git itself to treat a directory as a "real" .git
folder; and anyway if the HEAD file is absent Meson will fall
back to the "git rev-parse --git-dir" mechanism rather than
fail completely.

Fixes: #8567
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>